LRAC-4859 rename package
