+++ modules/apps/analytics/analytics-settings-api/bnd.bnd
@@ -3,4 +3,4 @@ Bundle-SymbolicName: com.liferay.analytics.settings.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.analytics.settings.configuration,\
-	com.liferay.analytics.settings.util
+	com.liferay.analytics.settings.user
