+++ modules/apps/analytics/analytics-settings-api/src/main/resources/com/liferay/analytics/settings/user/packageinfo
